Move physical host worker to own module
Adding the device worker to the same file is going to make this module a bit
too thorny.
