+++ doni/driver/worker/blazar/device.py
@@ -1,0 +1,15 @@
+from doni.driver.worker.base import BaseWorker
+from doni.worker import WorkerField
+
+
+class BlazarDeviceWorker(BaseWorker):
+    worker_fields = [
+        WorkerField(
+            "blazar_device_driver",
+            required=True,
+            description=(
+                "Which Blazar device driver plugin to use to make the device "
+                "reservable."
+            ),
+        )
+    ]
